Added try/except to catch weird data exception.
The exception is probably symptomatic of a bigger problem.  It is
thrown after the user complete a Quizly quiz.  That quiz's instance_id
is no longer valid as a key.  The exception causes a failure to load
student activity data.  The error seems to be this:
   Question or question-group ID "twxBgieSEwqs" in event is no longer present

The try/except allows that question to be skipped and the activity data to load.