Drops scores table from roster to improve efficiency.
This version will work for large class sizes.  For a class of 36
the roster loads in a few seconds.

What's dropped from roster view is the ability to student progres
on the quizzes in each lesson and to see an aggregate of the
class's project in a single table.  Too much data and/or processing
to make that work at this point.